ender3/skr-e3-dip: Initial SKR-E3-DIP Support
**NOTE: THIS DOES NOT PROPERLY ADD BLTOUCH SUPPORT YET**

Pulling unfucking some of the mis-mash of commits across various
@BIGTREETECH repositories (including
[BIGTREETECH-SKR-mini-E3-], [BIGTREETECH-SKR-E3-DIP-V1.0]).

Comparing this to a bunch of the changes in the bigtreetech repositories
you'll find a reversion of many of the changes to the comments while at
the same time selectively pulling in the initial changes they've made.
Were this not the equivalent of an "example" config, I would prefer to
have this broken down in a commit per grouping of features (e.g. "this
commit makes the necessary changes for SPI support on the stepper
motors") rather than one semi-opaque diff.
[BIGTREETECH-SKR-E3-DIP-V1.0]: https://github.com/bigtreetech/BIGTREETECH-SKR-E3-DIP-V1.0
[BIGTREETECH-SKR-mini-E3-]: https://github.com/bigtreetech/BIGTREETECH-SKR-mini-E3-